# KAFKA-10438 Lazy initialization of record header to reduce memory usa…