--- a/core/src/main/scala/kafka/log/LogValidator.scala
+++ b/core/src/main/scala/kafka/log/LogValidator.scala
@@ -77,7 +77,7 @@ private[log] object LogValidator extends Logging {
    * Returns a ValidationAndOffsetAssignResult containing the validated message set, maximum timestamp, the offset
    * of the shallow message with the max timestamp and a boolean indicating whether the message sizes may have changed.
    */
-  private[log] def validateMessagesAndAssignOffsets(records: MemoryRecords,
+  def validateMessagesAndAssignOffsets(records: MemoryRecords,
                                                     topicPartition: TopicPartition,
                                                     offsetCounter: LongRef,
                                                     time: Time,
@@ -339,7 +339,7 @@ private[log] object LogValidator extends Logging {
    * 2. When the target magic is not equal to batches' magic, meaning format conversion is needed.
    * 3. When the target magic is equal to V0, meaning absolute offsets need to be re-assigned.
    */
-  def validateMessagesAndAssignOffsetsCompressed(records: MemoryRecords,
+  private def validateMessagesAndAssignOffsetsCompressed(records: MemoryRecords,
                                                  topicPartition: TopicPartition,
                                                  offsetCounter: LongRef,
                                                  time: Time,