… clients/src/test/java/org/apache/kafka/common/record/DefaultRecordTest.java