@chia7712 2020-08-30
> Any thoughts on test for this?

The JMH is attached.