@chia7712 2020-09-02
the failed build (jdk8 + scala 2.12) is related to #8955