@chia7712 2020-09-11
```
kafka.network.DynamicConnectionQuotaTest.testDynamicListenerConnectionCreationRateQuota
```
pass on my local.