@chia7712 2020-09-14
```
 Build / JDK 8 / org.apache.kafka.connect.mirror.MirrorConnectorsIntegrationTest.testReplication	1 分 17 秒	1
 Build / JDK 15 / kafka.api.SaslSslAdminIntegrationTest.testCreateDeleteTopics	2 分 2 秒	1
 Build / JDK 15 / org.apache.kafka.streams.integration.EosBetaUpgradeIntegrationTest.shouldUpgradeFromEosAlphaToEosBeta[true]	1 分 21 秒	1
 Build / JDK 15 / org.apache.kafka.streams.integration.EosBetaUpgradeIntegrationTest.shouldUpgradeFromEosAlphaToEosBeta[true]	
```
I don't observe the relation between those failures and this patch.