@chia7712 2020-09-21
```
Build / JDK 8 / org.apache.kafka.streams.integration.EosBetaUpgradeIntegrationTest.shouldUpgradeFromEosAlphaToEosBeta[true]
```

only jdk 8 gets failure. This flaky is unrelated to this patch. @ijuma Please take a look.